Rearrange the files to conform to new the gem directory structure.
