+++ activerecord-userstamp.gemspec
@@ -1,7 +1,6 @@
 # -*- encoding: utf-8 -*-
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'active_record'
 require 'active_record/userstamp/version'
 
 Gem::Specification.new do |s|
@@ -20,6 +19,7 @@ Gem::Specification.new do |s|
   s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
   s.require_paths = ['lib']
 
+  s.add_dependency 'activesupport', '~> 4.2'
   s.add_dependency 'activerecord', '~> 4.2'
 
   s.add_development_dependency 'actionpack', '~> 4.2'
