+++ lib/active_record/userstamp.rb
@@ -1,0 +1,12 @@
+module ActiveRecord::Userstamp
+  extend ActiveSupport::Autoload
+
+  autoload :Stampable
+  autoload :Stamper
+
+  eager_autoload do
+    autoload :ControllerAdditions
+    autoload :MigrationAdditions
+    autoload :ModelAdditions
+  end
+end
