+++ lib/active_record/userstamp/version.rb
@@ -1,3 +1,4 @@
+module ActiveRecord; end
 module ActiveRecord::Userstamp
   VERSION = '4.0.0'
 end
