+++ lib/activerecord/userstamp.rb
@@ -1,0 +1,1 @@
+require 'active_record/userstamp'
